pyproject.toml
## Minimal, focused project configuration for ImageNet-64 PyTorch case study

[build-system]
build-backend = "setuptools.build_meta"
requires = [
  "setuptools>=75.0.0",
  "wheel>=0.44.0",
]

[project]
name = "case-study"
description = "ImageNet-64 Case Study: PyTorch implementation for image classification"
authors = [
    {name = "Shahin Mohammadi", email = "shahin.mohammadi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
dynamic = ["version"]

dependencies = [
  # Core DL stack (CPU by default; install CUDA builds externally if needed)
  "torch>=2.2,<2.5",
  "torchvision>=0.17,<0.20",
  "pytorch-lightning>=2.2,<2.5",
  # Numeric + utils
  "numpy>=1.24,<2.1",
  "scipy>=1.10,<1.13",
  "Pillow>=9.5",
  "PyYAML>=6.0",
  "tqdm>=4.65",
  "requests>=2.28",
  "matplotlib>=3.7",
  # PDF parsing for reading docs/description.pdf
  "pypdf>=4.2",
  # Clustering algorithms & metrics
  "scikit-learn>=1.3",
  # Single-cell data handling
  "anndata==0.9.2",
  "scanpy>=1.10",
]

[project.optional-dependencies]
dev = [
  # Testing
  "pytest>=8.0",
  "pytest-cov>=5.0",
  # Lint/format/type
  "ruff>=0.5",
  "black>=24.0",
  "mypy>=1.8",
  "pre-commit>=3.5",
  # Notebooks
  "jupyter>=1.0",
  "ipykernel>=6.29",
  "nbstripout>=0.7",
]

[tool.setuptools_scm]
write_to = "src/study/_version.py"



[tool.setuptools]
include-package-data = true
packages = ["study"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools-scm]
fallback_version = "0.1.0"
write_to = "src/study/_version.py"

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
